backend/requirements.txt
flask
flask_cors
PyYAML